pyproject.toml
[project]
name = "luks-unlock"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "python-pam",
    "six (>=1.17.0,<2.0.0)",
    "uwsgi (>=2.0.31,<3.0.0)",
]
